Remove AutoMapper

+++ src/Serval.Translation/Models/WordGraph.cs
@@ -2,7 +2,7 @@
 
 public class WordGraph
 {
-    public float InitialStateScore { get; set; }
+    public double InitialStateScore { get; set; }
     public List<int> FinalStates { get; set; } = new List<int>();
     public List<WordGraphArc> Arcs { get; set; } = new List<WordGraphArc>();
 }
